requirements.txt
python-telegram-bot>=22.0
aiohttp>=3.9.0
edge-tts>=7.0.0
soundfile>=0.13.0
numpy>=1.24.0
qwen-tts>=0.1.0
transformers>=4.57.0
accelerate>=1.10.0
safetensors>=0.4.0
librosa>=0.10.0
onnxruntime>=1.17.0
sox>=1.5.0
python-dotenv>=1.0.0